requirements.txt
openai
PyPDF2
pdf2image
pytesseract
Pillow
Flask
python-dotenv
